implement the adapter API for assert-activerecord adapter gems
This allows keeping the framework code in the assert-activerecord
gem and roll adapter gems for logic that is specific to each major
version of ActiveRecord.

This includes the adapter logic and tests, plus a README write up.

This also temporarily points the Gemfile at a dev feature branch
of AssertActiveRecord so CI will run.